+++ tests/functional_tests/test_run_query_bar.py
@@ -119,8 +119,14 @@ async def test_transaction_button(
         while app.editor is None or app.connection is None:
             await pilot.pause()
 
-        assert app.connection.transaction_mode == "Auto"
+        assert app.connection.transaction_mode
+        assert app.connection.transaction_mode.label == "Auto"
         snap_results.append(await app_snapshot(app, "Initialize with Tx: Auto"))
         await pilot.click("#transaction_button")
-        assert app.connection.transaction_mode == "Manual"
+        assert app.connection.transaction_mode
+        assert app.connection.transaction_mode.label == "Manual"
+        assert app.connection.transaction_mode.commit is not None
+        assert app.connection.transaction_mode.rollback is not None
         snap_results.append(await app_snapshot(app, "After click with Tx: Manual"))
+
+        assert all(snap_results)